+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Next
 - fixed RSuite radio-group missing required icon
 - fixed RSuite fullWidth / Width params in select
-- added MUI, RSuite Time
+- added time component for MUI, RSuite, Mantine
 - fixed required icon for MUI for all components
 - cleanup and more properties for MUI Date and MUI Datime